PP-4301: Move other create gateway account tests
@oswaldquek
