+++ src/test/java/uk/gov/pay/connector/it/resources/CreateGatewayAccountResourceTest.java
@@ -1,6 +1,8 @@
 package uk.gov.pay.connector.it.resources;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableMap;
+import com.jayway.restassured.response.ValidatableResponse;
 import junitparams.Parameters;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,8 +15,12 @@ import uk.gov.pay.connector.junit.DropwizardJUnitRunner;
 import static com.jayway.restassured.http.ContentType.JSON;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.text.IsEmptyString.isEmptyString;
 import static org.hamcrest.text.MatchesPattern.matchesPattern;
+import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity.Type.LIVE;
+import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity.Type.TEST;
+import static uk.gov.pay.connector.util.JsonEncoder.toJson;
 
 @RunWith(DropwizardJUnitRunner.class)
 @DropwizardConfig(app = ConnectorApp.class, config = "config/test-it-config.yaml")
@@ -64,4 +70,72 @@ public class CreateGatewayAccountResourceTest extends GatewayAccountResourceTest
     }
     
     //TODO get account /v1/frontend/accounts/{accountId} returns stripe credentials
+
+    @Test
+    public void createGatewayAccountWithoutPaymentProviderDefaultsToSandbox() {
+        String payload = toJson(ImmutableMap.of("name", "test account","type","test"));
+
+        ValidatableResponse response = givenSetup()
+                .body(payload)
+                .post(ACCOUNTS_API_URL)
+                .then()
+                .statusCode(201);
+
+        assertCorrectCreateResponse(response);
+        assertGettingAccountReturnsProviderName(response, "sandbox", TEST);
+    }
+
+    @Test
+    public void createGatewayAccount_shouldNotReturnCorporateSurcharges() {
+        String payload = toJson(ImmutableMap.of("name", "test account"));
+        ValidatableResponse response = givenSetup()
+                .body(payload)
+                .post(ACCOUNTS_API_URL)
+                .then()
+                .statusCode(201);
+
+        response.body("corporate_credit_card_surcharge_amount", is(nullValue()));
+        response.body("corporate_debit_card_surcharge_amount", is(nullValue()));
+        response.body("corporate_prepaid_credit_card_surcharge_amount", is(nullValue()));
+        response.body("corporate_prepaid_debit_card_surcharge_amount", is(nullValue()));
+    }
+
+    @Test
+    public void createGatewayAccountWithProviderUrlTypeLive() {
+        String payload = toJson(ImmutableMap.of("payment_provider", "worldpay", "type", LIVE.toString()));
+        ValidatableResponse response = givenSetup()
+                .body(payload)
+                .post(ACCOUNTS_API_URL)
+                .then()
+                .statusCode(201);
+
+        assertCorrectCreateResponse(response, LIVE);
+        assertGettingAccountReturnsProviderName(response, "worldpay", LIVE);
+    }
+
+    @Test
+    public void createGatewayAccountWithNameDescriptionAndAnalyticsId() {
+        String payload = toJson(ImmutableMap.of("service_name", "my service name", "description", "desc", "analytics_id", "analytics-id"));
+        ValidatableResponse response = givenSetup()
+                .body(payload)
+                .post(ACCOUNTS_API_URL)
+                .then()
+                .statusCode(201);
+
+        assertCorrectCreateResponse(response, TEST, "desc", "analytics-id", "my service name");
+        assertGettingAccountReturnsProviderName(response, "sandbox", TEST);
+    }
+
+    @Test
+    public void createGatewayAccountWithMissingProviderUrlTypeCreatesTestType() {
+        String payload = toJson(ImmutableMap.of("payment_provider", "worldpay"));
+        ValidatableResponse response = givenSetup()
+                .body(payload)
+                .post(ACCOUNTS_API_URL)
+                .then()
+                .statusCode(201);
+
+        assertCorrectCreateResponse(response, TEST);
+        assertGettingAccountReturnsProviderName(response, "worldpay", TEST);
+    }
 }
